Create Chatting Application
This project is a chat application developed in Java that uses the socket package for client-server communication and Java Swing for the graphical user interface. It allows real-time messaging between clients and a server, with the server managing multiple client connections and handling message distribution. The GUI, designed with Swing, provides a user-friendly interface for sending and receiving messages, complete with real-time timestamps to track communication. This project showcases the integration of network programming with GUI development to create a functional and interactive chat system.